xe: ocl: fix offset calculation in prelu

+++ src/gpu/intel/ocl/ref_prelu.cl
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright 2020-2024 Intel Corporation
+* Copyright 2020-2025 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -23,15 +23,15 @@
 __kernel void ref_prelu_fwd(const __global SRC_DATA_T *src,
         const __global WEI_DATA_T *weights, __global DST_DATA_T *dst) {
 
-    const int d0 = GWS_GET_D0();
-    const int d1 = GWS_GET_D1();
-    const int d2 = GWS_GET_D2();
-    const int d3 = GWS_GET_D3();
-    const int d4 = GWS_GET_D4();
-    const int d5 = GWS_GET_D5();
+    const off_t d0 = GWS_GET_D0();
+    const off_t d1 = GWS_GET_D1();
+    const off_t d2 = GWS_GET_D2();
+    const off_t d3 = GWS_GET_D3();
+    const off_t d4 = GWS_GET_D4();
+    const off_t d5 = GWS_GET_D5();
 
-    const unsigned data_off = OFF_MD(SRC, d0, d1, d2, d3, d4, d5);
-    const unsigned wei_off = OFF_MD(WEI, d0 % WEI_D0, d1 % WEI_D1, d2 % WEI_D2,
+    const off_t data_off = OFF_MD(SRC, d0, d1, d2, d3, d4, d5);
+    const off_t wei_off = OFF_MD(WEI, d0 % WEI_D0, d1 % WEI_D1, d2 % WEI_D2,
             d3 % WEI_D3, d4 % WEI_D4, d5 % WEI_D5);
 
     const float src_data = SRC_TO_REF(src[data_off]);
@@ -49,18 +49,18 @@ __kernel void ref_prelu_bwd(const __global SRC_DATA_T *src,
         const __global WEI_DATA_T *weights, const __global DST_DATA_T *diff_dst,
         __global DIFF_SRC_DATA_T *diff_src,
         __global DIFF_WEI_DATA_T *diff_weights) {
-    const int d0 = GWS_GET_D0();
-    const int d1 = GWS_GET_D1();
-    const int d2 = GWS_GET_D2();
-    const int d3 = GWS_GET_D3();
-    const int d4 = GWS_GET_D4();
-    const int d5 = GWS_GET_D5();
+    const off_t d0 = GWS_GET_D0();
+    const off_t d1 = GWS_GET_D1();
+    const off_t d2 = GWS_GET_D2();
+    const off_t d3 = GWS_GET_D3();
+    const off_t d4 = GWS_GET_D4();
+    const off_t d5 = GWS_GET_D5();
 
-    const unsigned data_off = OFF_MD(SRC, d0 % SRC_D0, d1 % SRC_D1, d2 % SRC_D2,
+    const off_t data_off = OFF_MD(SRC, d0 % SRC_D0, d1 % SRC_D1, d2 % SRC_D2,
             d3 % SRC_D3, d4 % SRC_D4, d5 % SRC_D5);
-    const unsigned data_off_pd = OFF_MD(SRC, d0 % SRC_PD0, d1 % SRC_PD1,
+    const off_t data_off_pd = OFF_MD(SRC, d0 % SRC_PD0, d1 % SRC_PD1,
             d2 % SRC_PD2, d3 % SRC_PD3, d4 % SRC_PD4, d5 % SRC_PD5);
-    const unsigned wei_off = OFF_MD(WEI, d0 % WEI_D0, d1 % WEI_D1, d2 % WEI_D2,
+    const off_t wei_off = OFF_MD(WEI, d0 % WEI_D0, d1 % WEI_D1, d2 % WEI_D2,
             d3 % WEI_D3, d4 % WEI_D4, d5 % WEI_D5);
 
     const float src_data = SRC_TO_REF(src[data_off]);
@@ -86,10 +86,10 @@ __kernel void ref_prelu_bwd(const __global SRC_DATA_T *src,
         diff_src[data_off_pd] = TO_SRC(diff_src_data);
     if (!COORDINATES_ARE_IN_RANGE(DIFF_WEI)) return;
 
-    const unsigned diff_wei_off = OFF_MD(DIFF_WEI, d0 % DIFF_WEI_D0,
+    const off_t diff_wei_off = OFF_MD(DIFF_WEI, d0 % DIFF_WEI_D0,
             d1 % DIFF_WEI_D1, d2 % DIFF_WEI_D2, d3 % DIFF_WEI_D3,
             d4 % DIFF_WEI_D4, d5 % DIFF_WEI_D5);
-    const unsigned diff_wei_off_pd = OFF_MD(DIFF_WEI, d0 % DIFF_WEI_PD0,
+    const off_t diff_wei_off_pd = OFF_MD(DIFF_WEI, d0 % DIFF_WEI_PD0,
             d1 % DIFF_WEI_PD1, d2 % DIFF_WEI_PD2, d3 % DIFF_WEI_PD3,
             d4 % DIFF_WEI_PD4, d5 % DIFF_WEI_PD5);
